requirements.txt
# Core framework:
fastapi

# Server to run the app:
uvicorn

# JWT security and authentication:
PyJWT

# Configuration management:
python-decouple

# Email validator:
pydantic[email]

# Encrypt passwords:
passlib
bcrypt

# Database:
mysql-connector-python